requirements.txt
absl-py==0.12.0
asgiref==3.3.1
astor==0.8.1
astunparse==1.6.3
bleach==1.5.0
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.7
django-heroku==0.3.1
django-mysql==3.11.1
djangorestframework==3.12.2
flatbuffers==1.12
gast==0.2.2
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
html5lib==0.9999999
idna==2.10
importlib-metadata==3.7.2
Keras==2.3.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy==1.19.5
oauthlib==3.1.0
opencv-contrib-python==4.5.1.48
opencv-python-headless==4.2.0.32
opt-einsum==3.3.0
Pillow==8.1.2
protobuf==3.15.6
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.25.1
PyJWT==2.0.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.5.4
six==1.15.0
scikit-image==0.18.0rc1
sqlparse==0.4.1
tensorboard==2.0.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.4.1